+++ douge/src/main/java/lv/javaguru/java2/domain/MetricSet.java
@@ -12,46 +12,46 @@ public class MetricSet {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name="id", nullable = false, columnDefinition = "INT(11)")
-    private Integer id;
+    private Long id;
 
     @Column(name="primary_id", nullable = false, columnDefinition = "INT(11)")
-    private Integer primary_id;
+    private Long primary_id;
 
     @Column(name="groupby_id", nullable = false, columnDefinition = "INT(11)")
-    private Integer groupby_id;
+    private Long groupby_id;
 
     @Column(name="limit_id", nullable = false, columnDefinition = "INT(11)")
-    private Integer limit_id;
+    private Long limit_id;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getPrimary_id() {
+    public Long getPrimary_id() {
         return primary_id;
     }
 
-    public void setPrimary_id(Integer primary_id) {
+    public void setPrimary_id(Long primary_id) {
         this.primary_id = primary_id;
     }
 
-    public Integer getGroupby_id() {
+    public Long getGroupby_id() {
         return groupby_id;
     }
 
-    public void setGroupby_id(Integer groupby_id) {
+    public void setGroupby_id(Long groupby_id) {
         this.groupby_id = groupby_id;
     }
 
-    public Integer getLimit_id() {
+    public Long getLimit_id() {
         return limit_id;
     }
 
-    public void setLimit_id(Integer limit_id) {
+    public void setLimit_id(Long limit_id) {
         this.limit_id = limit_id;
     }
 }
